renamed utils folder on client side utilsClient for debugging
