+++ client/src/components/QuestionList.jsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
-import { GET_QUESTIONS } from '../utils/queries/questionQueries';
+import { GET_QUESTIONS } from '../utilsClient/queries/questionQueries';
 import styled from 'styled-components';
 
 // Styled-components
