+++ client/src/pages/Login.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useMutation } from '@apollo/client';
-import { LOGIN_USER } from '../utils/mutations';
+import { LOGIN_USER } from '../utilsClient/mutations';
 
-import Auth from '../utils/auth';
+import Auth from '../utilsClient/auth';
 
 const Login = (props) => {
   const [formState, setFormState] = useState({ email: '', password: '' });
