+++ client/src/pages/SignUp.jsx
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 
 import { useMutation } from '@apollo/client';
-import { ADD_PROFILE } from '../utils/mutations';
+import { ADD_PROFILE } from '../utilsClient/mutations';
 
-import Auth from '../utils/auth';
+import Auth from '../utilsClient/auth';
 
 const SignUp = () => {
   const [formState, setFormState] = useState({
